ManifestUtilities

Documentation for ManifestUtilities.

ManifestUtilities.prune_manifest

ManifestUtilities.prune_manifest — Method

prune_manifest(io::IO; project_filename::AbstractString, manifest_filename::AbstractString)

Read the project and manifest files from the respective filenames, remove all packages from the manifest that are not indirect (recursive) dependencies of the project file, and print the new manifest to IO. The original project and manifest files are not modified.