ManifestUtilities

Documentation for ManifestUtilities.

ManifestUtilities.prune_manifest

ManifestUtilities.prune_manifest

ManifestUtilities.prune_manifest — Method

prune_manifest(io::IO; kwargs...)

Parse the given project and manifest, and generate a new manifest that only includes packages that are direct or indirect (recursive) dependencies of the given project. The new manifest is printed to the given IO. The original project and manifest are not modified.

Required Keyword Arguments

project_filename::AbstractString: the filename of the input Project.toml file

manifest_filename::AbstractString: the filename of the input Manifest.toml file

ManifestUtilities.prune_manifest — Method

prune_manifest(; kwargs...) -> AbstractString

Parse the given project and manifest, and generate a new manifest that only includes packages that are direct or indirect (recursive) dependencies of the given project. The new manifest is returned as an AbstractString. The original project and manifest are not modified.

Required Keyword Arguments

project_filename::AbstractString: the filename of the input Project.toml file

manifest_filename::AbstractString: the filename of the input Manifest.toml file